Move mattermost observer into observers.py
